fix(Backup): add healthcheck to full backup
